Websocket children restart strategy should be temporary.
With the current settings (none provided) the restart strategy is
permanent, this means a restart will be attempted in all cases.

For something as transient as a websocket this is bad because when
it is closed you'll go into a restart loop, you'll reach your max
restart limit and the supervisor will drag all the open connections
with it.